Move error handler interfaces to transport package
