--- transport/amqp/error_handler.go
@@ -1,7 +1,0 @@
-package amqp
-
-// ErrorHandler receives a transport error to be processed for diagnostic purposes.
-// Usually this means logging the error.
-type ErrorHandler interface {
-	Handle(err error)
-}
